requirements.txt
streamlit==1.23.1
requests
streamlit_option_menu
streamlit-authenticator
google-cloud-storage==2.9.0
SQLAlchemy
python-dotenv
cloud-sql-python-connector
cloud-sql-python-connector[pg8000]
requests
PyPDF2
google-cloud-storage
google-cloud-pubsub
pydantic